update mid debug version, temp work once.

+++ Core/app/Inc/bmp280.h
@@ -4,6 +4,8 @@
 #include "stm32f4xx_hal.h"
 
 void BMP280_Init(I2C_HandleTypeDef *hi2c);
-float BMP280_ReadTemperature(void);
+int32_t BMP280_ReadTemperature(void);
+
+
 
 #endif
